requirements.txt
functions-framework==3.*
google-cloud-secret-manager==2.20.0
google-api-python-client==2.120.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-genai==1.9.0
